Move the product list skeleton to the product list component folder
