+++ packages/js/product-editor/src/blocks/generic/linked-product-list/editor.scss
@@ -1,5 +1,3 @@
-@import "./skeleton/style.scss";
-
 .wp-block-woocommerce-product-linked-list-field {
 	display: flex;
 	flex-direction: column;
