+++ packages/js/product-editor/src/components/product-list/index.ts
@@ -1,2 +1,3 @@
 export * from './product-list';
 export * from './types';
+export * from './skeleton';
