+++ packages/js/product-editor/src/components/product-list/skeleton/index.ts
