+++ packages/js/product-editor/src/components/product-list/skeleton/style.scss
@@ -1,4 +1,4 @@
-.wp-block-woocommerce-product-linked-list-field {
+.woocommerce-product-list {
 	@mixin skeleton {
 		@include placeholder();
 		background-color: $gray-200;
@@ -11,6 +11,11 @@
 
 	[role="columnheader"] .skeleton {
 		width: $grid-unit-80;
+	}
+
+	&__product-image.skeleton {
+		width: $grid-unit-40;
+		height: $grid-unit-40;
 	}
 
 	&__product-name.skeleton {
