+++ packages/js/product-editor/src/components/product-list/styles.scss
@@ -1,3 +1,5 @@
+@import "./skeleton/style.scss";
+
 .woocommerce-product-list {
 	[role="table"] {
 		width: 100%;
